Use unsigned length in sds header
People have reported redis-cli segfaults because
of large sds allocations.

[Additional fixes by matt:
  - add the unsigned storage class to sds in hiredis
  - convert sdsIncrLen from int to long so assert works]

All tests pass.

Fixes #1188

+++ src/sds.c
@@ -197,7 +197,7 @@ size_t sdsAllocSize(sds s) {
  * ... check for nread <= 0 and handle it ...
  * sdsIncrLen(s, nread);
  */
-void sdsIncrLen(sds s, int incr) {
+void sdsIncrLen(sds s, long incr) {
     struct sdshdr *sh = (void*) (s-(sizeof(struct sdshdr)));
 
     assert(sh->free >= incr);
